pyproject.toml
[tool.poetry]
name = "pub-membres"
version = "0.1.0"
description = ""
authors = ["olichose123 <olivier@banville.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
